+++ server/src/main/java/org/opensearch/action/ActionModule.java
@@ -480,7 +480,6 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentSkipListSet;
@@ -492,7 +491,6 @@ import java.util.stream.Stream;
 
 import static java.util.Collections.unmodifiableMap;
 import static java.util.Objects.requireNonNull;
-import static org.opensearch.rest.RestController.PASS_THROUGH_REST_HANDLER_WRAPPER;
 
 /**
  * Builds and binds the generic action map, all {@link TransportAction}s, and {@link ActionFilters}.
@@ -524,7 +522,6 @@ public class ActionModule extends AbstractModule {
     private final AutoCreateIndex autoCreateIndex;
     private final DestructiveOperations destructiveOperations;
     private final RestController restController;
-    private final UnaryOperator<RestHandler> restWrapper;
     private final RequestValidators<PutMappingRequest> mappingRequestValidators;
     private final RequestValidators<IndicesAliasesRequest> indicesAliasesRequestRequestValidators;
     private final ThreadPool threadPool;
@@ -562,21 +559,17 @@ public class ActionModule extends AbstractModule {
             actionPlugins.stream().flatMap(p -> p.getRestHeaders().stream()),
             Stream.of(new RestHeaderDefinition(Task.X_OPAQUE_ID, false))
         ).collect(Collectors.toSet());
-        UnaryOperator<RestHandler> restWrapper = identityService.authenticate();
-        // Check if implementation is provided by one of the actionPlugins
-        if (PASS_THROUGH_REST_HANDLER_WRAPPER.equals(restWrapper)) {
-            List<UnaryOperator<RestHandler>> restWrappers = actionPlugins.stream()
-                .map(p -> p.getRestHandlerWrapper(threadPool.getThreadContext()))
-                .filter(Objects::nonNull)
-                .collect(Collectors.toUnmodifiableList());
-            if (!restWrappers.isEmpty()) {
-                if (restWrappers.size() > 1) {
+        UnaryOperator<RestHandler> restWrapper = null;
+        for (ActionPlugin plugin : actionPlugins) {
+            UnaryOperator<RestHandler> newRestWrapper = plugin.getRestHandlerWrapper(threadPool.getThreadContext());
+            if (newRestWrapper != null) {
+                logger.debug("Using REST wrapper from plugin " + plugin.getClass().getName());
+                if (restWrapper != null) {
                     throw new IllegalArgumentException("Cannot have more than one plugin implementing a REST wrapper");
                 }
-                restWrapper = restWrappers.get(0);
+                restWrapper = newRestWrapper;
             }
         }
-        this.restWrapper = restWrapper;
         mappingRequestValidators = new RequestValidators<>(
             actionPlugins.stream().flatMap(p -> p.mappingRequestValidators().stream()).collect(Collectors.toList())
         );
@@ -1064,11 +1057,6 @@ public class ActionModule extends AbstractModule {
 
     public RestController getRestController() {
         return restController;
-    }
-
-    // Visible for testing
-    UnaryOperator<RestHandler> getRestWrapper() {
-        return restWrapper;
     }
 
     /**