+++ server/src/main/java/org/opensearch/plugins/IdentityPlugin.java
@@ -9,15 +9,9 @@
 package org.opensearch.plugins;
 
 import org.opensearch.common.annotation.ExperimentalApi;
-import org.opensearch.common.util.concurrent.ThreadContext;
 import org.opensearch.identity.PluginSubject;
 import org.opensearch.identity.Subject;
 import org.opensearch.identity.tokens.TokenManager;
-import org.opensearch.rest.RestHandler;
-
-import java.util.function.UnaryOperator;
-
-import static org.opensearch.rest.RestController.PASS_THROUGH_REST_HANDLER_WRAPPER;
 
 /**
  * Plugin that provides identity and access control for OpenSearch
@@ -39,15 +33,6 @@ public interface IdentityPlugin {
      * @return Should never return null.
      */
     TokenManager getTokenManager();
-
-    /**
-     * Returns a function used to wrap each rest request before handling the request.
-     * The returned {@link UnaryOperator} is called for every incoming rest request and receives
-     * the original rest handler as it's input.
-     */
-    default UnaryOperator<RestHandler> authenticate(ThreadContext threadContext) {
-        return PASS_THROUGH_REST_HANDLER_WRAPPER;
-    }
 
     /**
      * Gets a subject corresponding to the passed plugin that can be utilized to perform transport actions