+++ server/src/test/java/org/opensearch/action/ActionModuleTests.java
@@ -49,13 +49,8 @@ import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.action.ActionResponse;
 import org.opensearch.extensions.ExtensionsManager;
 import org.opensearch.identity.IdentityService;
-import org.opensearch.identity.PluginSubject;
-import org.opensearch.identity.Subject;
-import org.opensearch.identity.tokens.TokenManager;
 import org.opensearch.plugins.ActionPlugin;
 import org.opensearch.plugins.ActionPlugin.ActionHandler;
-import org.opensearch.plugins.IdentityPlugin;
-import org.opensearch.plugins.Plugin;
 import org.opensearch.rest.RestChannel;
 import org.opensearch.rest.RestController;
 import org.opensearch.rest.RestHandler;
@@ -74,15 +69,10 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 import java.util.function.Supplier;
-import java.util.function.UnaryOperator;
 
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singletonList;
-import static org.opensearch.rest.RestController.PASS_THROUGH_REST_HANDLER_WRAPPER;
-import static org.hamcrest.Matchers.containsString;
-import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasEntry;
-import static org.hamcrest.Matchers.not;
 import static org.hamcrest.Matchers.startsWith;
 import static org.mockito.Mockito.mock;
 
@@ -171,124 +161,6 @@ public class ActionModuleTests extends OpenSearchTestCase {
             })
         );
         assertThat(e.getMessage(), startsWith("Cannot replace existing handler for [/] for method: GET"));
-        assertThat(actionModule.getRestWrapper(), equalTo(PASS_THROUGH_REST_HANDLER_WRAPPER));
-    }
-
-    private class TestIdentityPlugin implements IdentityPlugin {
-
-        @Override
-        public Subject getCurrentSubject() {
-            return null;
-        }
-
-        @Override
-        public TokenManager getTokenManager() {
-            return null;
-        }
-
-        @Override
-        public UnaryOperator<RestHandler> authenticate(ThreadContext threadContext) {
-            return (h) -> {
-                threadContext.putHeader("test_header", "foo");
-                return h;
-            };
-        }
-
-        @Override
-        public PluginSubject getPluginSubject(Plugin plugin) {
-            return null;
-        }
-    }
-
-    private class TestActionPlugin implements ActionPlugin {
-
-        private ThreadPool threadPool;
-
-        public TestActionPlugin(ThreadPool threadPool) {
-            this.threadPool = threadPool;
-        }
-
-        @Override
-        public UnaryOperator<RestHandler> getRestHandlerWrapper(ThreadContext threadContext) {
-            return (h) -> {
-                threadContext.putHeader("test_header", "bar");
-                return h;
-            };
-        }
-    }
-
-    public void testSetupRestHandlerWithIdentityPlugin() throws IOException {
-        ThreadPool threadPool = mock(ThreadPool.class);
-        IdentityPlugin testIdentityPlugin = new TestIdentityPlugin();
-        SettingsModule settings = new SettingsModule(Settings.EMPTY);
-        UsageService usageService = new UsageService();
-        IdentityService identityService = new IdentityService(Settings.EMPTY, mock(ThreadPool.class), List.of(testIdentityPlugin));
-        ActionModule actionModule = new ActionModule(
-            settings.getSettings(),
-            new IndexNameExpressionResolver(new ThreadContext(Settings.EMPTY)),
-            settings.getIndexScopedSettings(),
-            settings.getClusterSettings(),
-            settings.getSettingsFilter(),
-            threadPool,
-            emptyList(),
-            null,
-            null,
-            usageService,
-            null,
-            identityService,
-            new ExtensionsManager(Set.of(), identityService)
-        );
-        assertThat(actionModule.getRestWrapper(), not(equalTo(PASS_THROUGH_REST_HANDLER_WRAPPER)));
-    }
-
-    public void testSetupRestHandlerWithActionPluginNoIdentityPlugin() throws IOException {
-        ThreadPool threadPool = mock(ThreadPool.class);
-        TestActionPlugin testActionPlugin = new TestActionPlugin(threadPool);
-        SettingsModule settings = new SettingsModule(Settings.EMPTY);
-        UsageService usageService = new UsageService();
-        ActionModule actionModule = new ActionModule(
-            settings.getSettings(),
-            new IndexNameExpressionResolver(new ThreadContext(Settings.EMPTY)),
-            settings.getIndexScopedSettings(),
-            settings.getClusterSettings(),
-            settings.getSettingsFilter(),
-            threadPool,
-            List.of(testActionPlugin),
-            null,
-            null,
-            usageService,
-            null,
-            new IdentityService(Settings.EMPTY, mock(ThreadPool.class), List.of()),
-            new ExtensionsManager(Set.of(), new IdentityService(Settings.EMPTY, mock(ThreadPool.class), List.of()))
-        );
-        assertThat(actionModule.getRestWrapper(), not(equalTo(PASS_THROUGH_REST_HANDLER_WRAPPER)));
-    }
-
-    public void testSetupRestHandlerWithIdentityPluginOverrideActionPlugin() throws IOException {
-        IdentityPlugin testIdentityPlugin = new TestIdentityPlugin();
-        ThreadPool threadPool = mock(ThreadPool.class);
-        TestActionPlugin testActionPlugin = new TestActionPlugin(threadPool);
-        SettingsModule settings = new SettingsModule(Settings.EMPTY);
-        UsageService usageService = new UsageService();
-        IdentityService identityService = new IdentityService(Settings.EMPTY, mock(ThreadPool.class), List.of(testIdentityPlugin));
-        ActionModule actionModule = new ActionModule(
-            settings.getSettings(),
-            new IndexNameExpressionResolver(new ThreadContext(Settings.EMPTY)),
-            settings.getIndexScopedSettings(),
-            settings.getClusterSettings(),
-            settings.getSettingsFilter(),
-            threadPool,
-            List.of(testActionPlugin),
-            null,
-            null,
-            usageService,
-            null,
-            identityService,
-            new ExtensionsManager(Set.of(), identityService)
-        );
-        assertThat(actionModule.getRestWrapper(), not(equalTo(PASS_THROUGH_REST_HANDLER_WRAPPER)));
-        assertThat(actionModule.getRestWrapper().getClass().getName(), not(containsString("TestActionPlugin")));
-        assertThat(actionModule.getRestWrapper().getClass().getName(), containsString("TestIdentityPlugin"));
     }
 
     public void testPluginCantOverwriteBuiltinRestHandler() throws IOException {